requirements.txt
flask==3.0.2
flask-cors==4.0.0
langchain==0.1.5
langchain-groq==0.0.1
python-dotenv==1.0.1
peewee==3.17.1
